Replace `FormSliderBar.Instantaneous` with `TransferValueOnCommit`
Rather than control the propagation of the value between the slider and
the textbox, add a property that controls the propagation of the value
between the bindables inside the form control to external bindables.
This will help alleviate issues where the external bindable update
incurs overheads due to having heavy change callbacks attached.