Return true while in gameplay
A `false` value marks the user as being on the latest release, and
notifies them as such when clicking the button in settings. In reality,
we don't know whether this is the case yet - we're just deferring the
check.

Somewhat minor change because the chance of a user manually going into
settings and clicking the button is very small.