Remove no longer required nested object reposition hacks
As touched on in
https://github.com/ppy/osu/issues/30237#issuecomment-2408557766, these
types of maneouvers are no longer required after
https://github.com/ppy/osu/pull/30021 - although as it turns out on
closer inspection, these things being there still *did not actually
break anything*, because the `slider.Path` mutation at the end of
`modifySlider()` causes `updateNestedPositions()` to be called
eventually anyway. So this is at mostly a code quality upgrade.