Fix strict tracking mod not populating path progress for ticks/repeats
Closes https://github.com/ppy/osu/issues/30237.

This is the root failure in the issue, and one that *only* presents when
another conversion mod that repositions the objects is also active. That
makes the `PathProgress` of the nesteds to be zero, therefore making
them occupy the position of the slider head after any mutation to the
path.