Rename setting to be more descriptive
this is honestly a nitpick on my behalf, but i figured it's better than
what we had before.

when i was looking at the setting in the settings menu, simulating a new
player, i asked myself "normalize audio? what audio?". for all the user
knows, this could be effects, beatmaps, hitsounds, anything.

this commit changes the text of the setting to read "Normalize Beatmap
Audio" to further describe that it's normalizing only beatmap audio