replace QueueUpdateHitObject with explicit set in transaction
This command doesn't make much sense as a "revertible change" and it requires assumptions about the debouncing of Update calls.
Instead I made a set in the Transaction that just stores the hit objects to update on undo/redo. I think this is pretty ugly and I'd prefer it if the `Apply()` of a change would invoke Update on its own, but I'm afraid that this might update the object too frequently and break drag behaviours.