Add search textbox in friends display
Random user feature request that made sense to me because the same thing
is in currently online display.

Yes web doesn't have this, but web is in a browser where you can Ctrl-F.
You can't do that in the client.

Design taken out of posterior because I can't be bothered waiting for
design cycles for this.