Add test case covering stability of compatibility export operation
This is important as the format will be used more when lazer beatmap
submission comes online, and its stability is a useful property for
that.

Included archive contains an `.osu` with a few fractional-millisecond
timing points and objects, as well as a multi-segment-type slider. That
should cover the range of all possible modifications that the
compatibility exporter currently performs.