Remove unnecessary local subscription in `BeatmapCarousel`
Not sure why I left this around during the refactor. This is 100%
handled by the `DetachedBeatmapStore`.

Removing this subscription reduces overheads by a huge amount for users
with large beatmap databases. My hypothesis is that subscriptions are
more expensive based on **the number of results matching**. This one
matches almost every beatmap so removing it is a large win.