Rename SkinInfo when it is changed in skin.ini
Peppy spoke about using a shortcut and/or hashes to determine if the
skin.ini is changed, and if so, then to rename the skin.

In my opinion, hashing and doing numerous comparisons is probably less
efficient than just syncing the SkinInfo's name during the update.

This is an easy solution that does what it needs to.