Display storyboard in editor background
Fixes the main part of https://github.com/ppy/osu/issues/31144.
Support for selecting a video will come later.

Making this work was an absolutely awful time full of dealing with
delightfully kooky issues, and yielded in a very weird-shaped
contraption. There is at least one issue remaining wherein storyboard
videos do not actually display until the track is started in editor, but
that is 99% a framework issue and I do not currently have the mental
fortitude to diagnose further.