Remove `Scheduler.AddOnce` from `updateSpecifics`
To keep things simple, let's not bother debouncing this. The debouncing
was causing spectating handling to fail because of two interdependent
components binding to `BeatmapAvailability`:

Binding to update the screen's `Beatmap` after a download completes:

https://github.com/ppy/osu/blob/58747061171c4ebe70201dfe4d3329ed7f4343f5/osu.Game/Screens/OnlinePlay/Match/RoomSubScreen.cs#L266-L267

Binding to attempt a load request:

https://github.com/ppy/osu/blob/b58ba5515da1904b3cce2d1f8a0bb038ab064345/osu.Game/Screens/OnlinePlay/Multiplayer/MultiplayerMatchSubScreen.cs#L67

The first must update the beatmap before the second runs, else gameplay
will not load due to `Beatmap.IsDefault`.