Adjust timeouts to be much higher for upload requests
It seems that right now these timeouts do not check for actual data
movement, which is to say if a user with a very slow connection is
uploading and it takes more than `Timeout`, their upload will fail.

For now let's set these values high enough that most users will not be
affected.