Fix score position not being displayed in solo results screen
Closes https://github.com/ppy/osu/issues/31842.

To be honest, I recall this working too, but I don't recall when it
might have broken, nor do I want to go look for the point of breakage
because it might be borderline impossible to find it now. So I'm just
fixing as if it was just a straight omission.

Opting for a client-side fix because server-side inclusion of the score
position for an entire leaderboard has been previously rejected as too
expensive:

	https://github.com/ppy/osu-web/pull/11354#discussion_r1689217450