Fix android build again
Another month, another freak android build failure.

From what I can tell, this time the build is broken because the second-
-to-last workaround applied to fix it, namely explicitly specifying
the version of workloads to install, stopped working, presumably because
github pushed a new .NET SDK version to runners, and microsoft didn't
actually put up a set of workload packages whose version matches the
.NET SDK version 1:1.

Thankfully, the fix to the *last* android build breakage (which caused
the workload installation to completely and irrecoverably fail), appears
to be rolling out this week, and *also* fix that same second-last issue,
so both workarounds of specifying the workload version and pinning the
image to `windows-2019` appear to no longer be required.

Note that the newest image version, 20250209.1.0, is still not fully
rolled out yet, thus rather than just fix all builds, this will fix like
20% of builds until the newest image is fully rolled out. But I guess
a 20% passing build is better than a 0% passing build, in a sense...?