Fix distance snap grid colours being off-by-one in certain cases
Closes https://github.com/ppy/osu/issues/31909.

Previously: https://github.com/ppy/osu/pull/30062.

Happening because of rounding errors - in this case the beat index
pre-flooring was something like a 0.003 off of a full beat, which would
get floored down rather than rounded up which created the discrepancy.
But also we don't want to round *too* far, which is why this
frankenstein solution has to exist I think. This is probably all
exacerbated by stable not handling decimal control point start times.

Would add tests if not for the fact that this is like extremely annoying
to test.