Attempt to fix spectator list showing other users in multiplayer room even if they're not spectating better
Maybe closes https://github.com/ppy/osu/issues/31972. Not sure. I have
no reproduction scenario to work with, no solid understanding of how
the issue can happen, and if this doesn't fix it, then I'm not even
entirely sure how this can ever be fixed client-side.

The working theory is that not watching updates to the room provoked a
situation wherein the room was temporarily not in a correct state when
`WatchingUsers` changed, therefore the collection change callback failed
to exclude other players in the room from display.

I'm only PRing this because of the `next-release` tag on the issue.