Do not give swell ticks any visual representation
Why is this a thing at all? How has it survived this long? I don't know.

As far as I can tell this only manifests on selected beatmaps with "slow
swells" that spend the entire beatmap moving in the background. On other
beatmaps the tick is faded out, probably due to the initial transform
application that normally "works" but fails hard on these slow swells.

Can be seen on https://osu.ppy.sh/beatmapsets/1432454#taiko/2948222.